Refactoring out of jeweler.rb
