+++ lib/bio-gem.rb
@@ -1,3 +1,5 @@
 #load Jeweler's mod for adapting it to our BioGem Project
 require 'bio-gem/mod/jeweler'
+require 'bio-gem/mod/biogem'
+require 'bio-gem/mod/biogem-rails'
 require 'bio-gem/application'
