+++ lib/bio-gem/mod/biogem-rails.rb
@@ -1,0 +1,14 @@
+# Using Biogem with Rails
+
+class String
+  # Handle underscore in routing template
+  def underscore
+    self.gsub(/::/, '/').
+    gsub(/([A-Z]+)([A-Z][a-z])/,'\1_\2').
+    gsub(/([a-z\d])([A-Z])/,'\1_\2').
+    tr("-", "_").
+    downcase
+  end
+end
+
+
